Moved Filter Token code
